+ phase 5.3 ablation #2 (no-TabPFN) + #5 (k_tabm=1) configs + bf16-numpy fix
bf16-numpy fix: on_validation_epoch_end now casts preds/targets to .float()
before .numpy() call. Required for Phase-1 fallback path (no-TabPFN ablation)
where the composite TabPFN-residual addition isn't there to promote bf16
predictions to fp32. Old code crashed with "TypeError: Got unsupported ScalarType BFloat16".

+++ src/training/resdec_lightning_module.py
@@ -469,9 +469,12 @@ class ResDecLightningModule(pl.LightningModule):
 
         # Pearson r (linear correlation) + Spearman ρ (rank correlation)
         # via numpy-on-CPU — simpler than torch-corrcoef broadcasting.
+        # .float() cast: numpy doesn't support bf16; under bf16-mixed precision
+        # the no-TabPFN-fallback path (Phase-1 plain MSE) leaves preds as bf16
+        # because there's no float-promotion via TabPFN-residual addition.
         import numpy as _np
-        p_np = preds.detach().numpy()
-        t_np = targets.detach().numpy()
+        p_np = preds.detach().float().numpy()
+        t_np = targets.detach().float().numpy()
         if p_np.std() > 0 and t_np.std() > 0:
             pearson_r = float(_np.corrcoef(p_np, t_np)[0, 1])
             # Spearman via rank-corr on np.argsort orderings
